requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.4
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
pycparser==2.22
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.4.4
